src/penroselamarck/orchestrator/python-requirements.txt
langgraph==0.2.75
langchain-openai==0.2.14
langfuse>=2,<3
opentelemetry-api==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-exporter-otlp==1.27.0
pydantic==2.11.9
SQLAlchemy==2.0.34
psycopg[binary]>=3.1,<4
